--- a/test/auto_clear_cache.uts
+++ b/test/auto_clear_cache.uts
@@ -7,7 +7,7 @@
 ############
 + Test utils
 
-= Prepare a `check_equal()` function.
+= 000) Prepare a `check_equal()` function.
 
 def check_equal(description, measured, expected):
     assert measured == expected, f"{description!r} FAILED: {measured!r} != {expected!r}"
@@ -18,7 +18,7 @@ def check_equal(description, measured, expected):
 ############
 + Test data
 
-= Define a final packet class.
+= 001) Define a final packet class.
 
 class F(Packet):
     fields_desc = [
@@ -30,27 +30,27 @@ class F(Packet):
 ############
 + Auto clear cache on final packet
 
-= Instantiate a packet of the final packet class with cache.
+= 002) Instantiate a packet of the final packet class with cache.
 
 f = F(bytes.fromhex("01"))
 f.show()
 check_equal("Cache is set", f.raw_packet_cache, bytes.fromhex("01"))
 
-= Change the value of the `F.value` field.
+= 003) Change the value of the `F.value` field.
 
 f.value = 2
 f.show()
 
-= Check the final packet has no cache anymore.
+= 004) Check the final packet has no cache anymore.
 
 check_equal("Cache is reset", f.raw_packet_cache, None)
 
-= Serialize the resulting packet.
+= 005) Serialize the resulting packet.
 
 buf = f.build()
 check_equal("f.build()", buf, bytes.fromhex("02"))
 
-= Check whether the final packet has cache after `build()`.
+= 006) Check whether the final packet has cache after `build()`.
 
 # Just watch, don't assert.
 # Note: Behaviour changed after #4705.
@@ -61,7 +61,7 @@ print(f"f.raw_packet_cache={f.raw_packet_cache!r}")
 ############
 + Auto clear cache with payload
 
-= Define a packet class bound with final class after it.
+= 007) Define a packet class bound with final class after it.
 
 class M0(Packet):
     fields_desc = [
@@ -70,28 +70,28 @@ class M0(Packet):
 
 bind_layers(M0, F)
 
-= Instantiate a packet of this class with cache.
+= 008) Instantiate a packet of this class with cache.
 
 m = M0(bytes.fromhex("01"))
 m.show()
 check_equal("Cache is set", m.raw_packet_cache, b'')
 check_equal("Cache is set", m.payload.raw_packet_cache, bytes.fromhex("01"))
 
-= Change the value of the `F.value` field.
+= 009) Change the value of the `F.value` field.
 
 m.value = 2
 m.show()
 
-= Check the underlayer packet has no cache anymore.
+= 010) Check the underlayer packet has no cache anymore.
 
 check_equal("Cache is reset", m.raw_packet_cache, None)
 
-= Serialize the resulting packet.
+= 011) Serialize the resulting packet.
 
 buf = m.build()
 check_equal("m.build()", buf, bytes.fromhex("02"))
 
-= Check whether the underlayer packet has cache.
+= 012) Check whether the underlayer packet has cache.
 
 # Just watch, don't assert.
 # Note: Behaviour changed after #4705.
@@ -102,34 +102,34 @@ print(f"m.raw_packet_cache={m.raw_packet_cache!r}")
 ############
 + Auto clear cache with `PacketField`
 
-= Define a packet class using a `PacketField`
+= 013) Define a packet class using a `PacketField`
 
 class M1(Packet):
     fields_desc = [
         PacketField("f", pkt_cls=F, default=F()),
     ]
 
-= Instantiate a packet of this class with cache.
+= 014) Instantiate a packet of this class with cache.
 
 m = M1(bytes.fromhex("01"))
 m.show()
 check_equal("Cache is set", m.raw_packet_cache, bytes.fromhex("01"))
 
-= Change the value of the `F.value` field.
+= 015) Change the value of the `F.value` field.
 
 m.f.value = 2
 m.show()
 
-= Check the parent packet has no cache anymore.
+= 016) Check the parent packet has no cache anymore.
 
 check_equal("Cache is reset", m.raw_packet_cache, None)
 
-= Serialize the resulting packet.
+= 017) Serialize the resulting packet.
 
 buf = m.build()
 check_equal("m.build()", buf, bytes.fromhex("02"))
 
-= Check whether the parent packet has cache.
+= 018) Check whether the parent packet has cache.
 
 # Just watch, don't assert.
 # Note: Behaviour changed after #4705.
@@ -140,34 +140,34 @@ print(f"m.raw_packet_cache={m.raw_packet_cache!r}")
 ############
 + Auto clear cache with `PacketListField`
 
-= Define a packet class using a `PacketListField`
+= 019) Define a packet class using a `PacketListField`
 
 class M2(Packet):
     fields_desc = [
         PacketListField("f", pkt_cls=F, count_from=lambda pkkt: 1, default=[]),
     ]
 
-= Instantiate a packet of this class with cache.
+= 020) Instantiate a packet of this class with cache.
 
 m = M2(bytes.fromhex("01"))
 m.show()
 check_equal("Cache is set", m.raw_packet_cache, bytes.fromhex("01"))
 
-= Change the value of the `F.value` field.
+= 021) Change the value of the `F.value` field.
 
 m.f[0].value = 2
 m.show()
 
-= Check the parent packet has no cache anymore.
+= 022) Check the parent packet has no cache anymore.
 
 check_equal("Cache is reset", m.raw_packet_cache, None)
 
-= Serialize the resulting packet.
+= 023) Serialize the resulting packet.
 
 buf = m.build()
 check_equal("m.build()", buf, bytes.fromhex("02"))
 
-= Check whether the parent packet has cache.
+= 024) Check whether the parent packet has cache.
 
 # Just watch, don't assert.
 # Note: Not reset in v2.4.5, reset in v2.5.0.