Draft PR for the following issues:
- #4705
- #4706
- #4707

Initiated as a draft PR for preliminary discussions before an official delivery (not squashed yet).

To be discussed:
- Changes currently based on v2.6.1 => shall I align on master? or useful to deliver a fix for a version before?
- `Packet.explicit` flag.
- `Packet.raw_packet_cache_fields` removed.

**Checklist:**

-   [x] If you are new to Scapy: I have checked [CONTRIBUTING.md](https://github.com/secdev/scapy/blob/master/CONTRIBUTING.md) (esp. section submitting-pull-requests)
-   [ ] I squashed commits belonging together
-   [x] I added unit tests or explained why they are not relevant
-   [ ] I executed the regression tests (using `cd test && ./run_tests` or `tox`)
-   [x] If the PR is still not finished, please create a [Draft Pull Request](https://github.blog/2019-02-14-introducing-draft-pull-requests/)