@alxroyer-thales 2025-04-23
A couple of figures about the impact of the changes:

'perf_packet_fields.uts' times before changes:
```
passed 2E618233 000.02s 000) Define `NUMBER_OF_I_PER_M` and `NUMBER_OF_F_PER_I` constants, and prepare a base `TestPacket` class.
passed 9DC21D20 000.00s 001) Define the `F` final packet class.
passed B7BBB574 000.01s 002) Define the `I` intermediate packet class.
passed E12AE3F0 000.54s 003) Define the `M` main packet class.
passed 10E80BEA 000.00s 004) Build a default instance of `F`.
passed 59DBEF73 000.01s 005) Build a default instance of `I`.
failed B573553A 002.91s 006) Build a default instance of `M`.
failed A6E98C89 005.31s 007) Launch serialization from the latest instance of `M` created.
failed 4DAB5BC5 005.06s 008) Launch serialization again from the same instance of `M`.
failed 46AD5D98 005.21s 009) Update one `F` from the same instance of `M` and launch serialization again.
passed B8CC085E 001.28s 010) Parse the buffer serialized before.
```

After changes:
```
passed 2E618233 000.02s 000) Define `NUMBER_OF_I_PER_M` and `NUMBER_OF_F_PER_I` constants, and prepare a base `TestPacket` class.
passed 9DC21D20 000.00s 001) Define the `F` final packet class.
passed B7BBB574 000.01s 002) Define the `I` intermediate packet class.
passed E12AE3F0 000.50s 003) Define the `M` main packet class.
passed 10E80BEA 000.00s 004) Build a default instance of `F`.
passed 59DBEF73 000.00s 005) Build a default instance of `I`.
passed B573553A 000.53s 006) Build a default instance of `M`.
passed A6E98C89 000.13s 007) Launch serialization from the latest instance of `M` created.
passed 4DAB5BC5 000.00s 008) Launch serialization again from the same instance of `M`.
passed 46AD5D98 000.00s 009) Update one `F` from the same instance of `M` and launch serialization again.
passed B8CC085E 001.22s 010) Parse the buffer serialized before.
```

Biggest impacts on default instantiations (fewer copies) and serializations (cache optimizations).
(It seems parsing had already been optimized with version 2.5.0.)

Before changes:
```
######
## Default instantiations
######


###(006)=[failed] 006) Build a default instance of `M`.

>>> TestPacket.begin_case()
>>> m = M()
>>> TestPacket.end_case(
...     init_m=1,
...     init_i=NUMBER_OF_I_PER_M,
...     init_f=NUMBER_OF_I_PER_M * NUMBER_OF_F_PER_I,
... )
Number of M inits: 1 (expected: 1)
Number of I inits: 200 (expected: 100)
Number of F inits: 60000 (expected: 10000)
Number of M builds: 0 (expected: 0)
Number of I builds: 0 (expected: 0)
Number of F builds: 0 (expected: 0)
Total number of operations expected: 10101
Execution time: 2.912 seconds (max: 1.667)
Traceback (most recent call last):
  File "<input>", line 5, in <module>
  File "<input>", line 45, in end_case
AssertionError: Execution time FAILED: 2.912 > 1.667


######
## Serialization
######


###(007)=[failed] 007) Launch serialization from the latest instance of `M` created.

>>> TestPacket.begin_case()
>>> buf = m.build()
>>> TestPacket.end_case(
...     build_m=1,
...     build_i=NUMBER_OF_I_PER_M,
...     build_f=NUMBER_OF_I_PER_M * NUMBER_OF_F_PER_I,
... )
Number of M inits: 1 (expected: 0)
Number of I inits: 300 (expected: 0)
Number of F inits: 90000 (expected: 0)
Number of M builds: 1 (expected: 1)
Number of I builds: 100 (expected: 100)
Number of F builds: 10000 (expected: 10000)
Total number of operations expected: 10101
Execution time: 5.307 seconds (max: 1.667)
Traceback (most recent call last):
  File "<input>", line 5, in <module>
  File "<input>", line 45, in end_case
AssertionError: Execution time FAILED: 5.307 > 1.667


###(008)=[failed] 008) Launch serialization again from the same instance of `M`.

>>> TestPacket.begin_case()
>>> buf = m.build()
>>> TestPacket.end_case(
...     build_m=1,  # `m` is cached, `build()` not spreaded on `i{n}` fields (thus, no `f{n}`).
...     build_i=0,
...     build_f=0,
... )
Number of M inits: 1 (expected: 0)
Number of I inits: 300 (expected: 0)
Number of F inits: 90000 (expected: 0)
Number of M builds: 1 (expected: 1)
Number of I builds: 100 (expected: 0)
Number of F builds: 10000 (expected: 0)
Total number of operations expected: 1
Execution time: 5.055 seconds (max: 0.010)
Traceback (most recent call last):
  File "<input>", line 5, in <module>
  File "<input>", line 45, in end_case
AssertionError: Execution time FAILED: 5.055 > 0.010


###(009)=[failed] 009) Update one `F` from the same instance of `M` and launch serialization again.

>>> m.i0.f0.value += 1
>>> 
>>> TestPacket.begin_case()
>>> buf = m.build()
>>> TestPacket.end_case(
...     build_m=1,
...     build_i=NUMBER_OF_I_PER_M,  # `i0` gets rebuilts, next `i{n}` fields are just asked for their cache.
...     build_f=1 * NUMBER_OF_F_PER_I,  # `i0.f0` gets rebuilt, next `i0.f{n}` fields are just asked for their cache.
... )
Number of M inits: 1 (expected: 0)
Number of I inits: 300 (expected: 0)
Number of F inits: 90000 (expected: 0)
Number of M builds: 1 (expected: 1)
Number of I builds: 100 (expected: 100)
Number of F builds: 10000 (expected: 100)
Total number of operations expected: 201
Execution time: 5.207 seconds (max: 0.033)
Traceback (most recent call last):
  File "<input>", line 5, in <module>
  File "<input>", line 45, in end_case
AssertionError: Execution time FAILED: 5.207 > 0.033
```