@alxroyer-thales 2025-04-24
I've also tried to see how things behave when making the two constants `NUMBER_OF_I_PER_M` and `NUMBER_OF_F_PER_I` vary.

In red, before the changes, in blue, after the changes:
![Figure_1](https://github.com/user-attachments/assets/5b8250ae-a0dc-4356-925d-f2791cbda2f9)

With that angle, we better see how the blue surface remains below the red one while `NUMBER_OF_I_PER_M` and `NUMBER_OF_F_PER_I` rise up.
![Figure_2](https://github.com/user-attachments/assets/e4839e35-a326-4853-9393-020a1e614a40)