@alxroyer-thales 2025-04-24
The code is probably rough for the moment. That's why I opened a draft PR.
I'll be glad to read through your remarks to make it better.