@alxroyer-thales 2025-06-27
> Hi, could you please fix the unit tests for your proposal. After that, I would go into a deeper review.

I'm on it.

I've written this [unittests-manual-utscapy.yml](https://github.com/ThalesGroup/scapy/blob/manual-ci/.github/workflows/unittests-manual-utscapy.yml) workflow file in order to be able to launch a single `utscapy` configuration from the matrix in 'unittests.yml'.

Good idea? or was there a simpler way to do it?